# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "Stable machine-readable code",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "description": "Allow-listed safe context",
            "title": "Details",
            "type": "object"
          },
          "message": {
            "description": "Safe human-readable text",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Server-generated request UUID (also in X-Request-ID)",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ModelStatus": {
        "description": "Sanitized model/policy readiness. No artifact paths, training identities or patient data.\n\n`model_loaded` alone is not `forecast_ready`: a verified bundle, its active calibrated policy and\na compatible MODEL_MODE are all required.",
        "properties": {
          "forecast_ready": {
            "title": "Forecast Ready",
            "type": "boolean"
          },
          "model_kind": {
            "anyOf": [
              {
                "enum": [
                  "LAST_VALUE",
                  "LINEAR_TREND",
                  "GRU"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Kind"
          },
          "model_loaded": {
            "title": "Model Loaded",
            "type": "boolean"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "policy_ready": {
            "title": "Policy Ready",
            "type": "boolean"
          },
          "policy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Sanitized reason when forecasting is not ready; null when ready",
            "title": "Reason Code"
          }
        },
        "required": [
          "model_kind",
          "model_version",
          "policy_version",
          "model_loaded",
          "policy_ready",
          "forecast_ready",
          "reason_code"
        ],
        "title": "ModelStatus",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Bearer token from POST /auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Cognitive monitoring prototype API (synthetic/demo data only; not a diagnostic or emergency service). All errors use the shared `ErrorResponse` envelope.",
    "title": "COGNUANCE",
    "version": "backend_contract_v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/model/status": {
      "get": {
        "description": "Model unavailability is a valid 200 document; 503 only when the status cannot be served.",
        "operationId": "getModelStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Infrastructure unavailable"
          }
        },
        "summary": "Model Status",
        "tags": [
          "health"
        ]
      }
    },
...
  }
}